+++ mininterface/interfaces.py
@@ -1,5 +1,8 @@
 # Access to interfaces via this module assures lazy loading
 from importlib import import_module
+import sys
+
+from .mininterface import Mininterface
 from .exceptions import InterfaceNotAvailable
 from .text_interface import TextInterface
 
@@ -9,6 +12,7 @@ def __getattr__(name):
     if name == "GuiInterface":
         return __getattr__("TkInterface")
     if name == "TuiInterface":
+        # if textual not installed or isatty False, return TextInterface
         return __getattr__("TextualInterface") or TextInterface
 
     # real interfaces
@@ -30,17 +34,26 @@ def __getattr__(name):
 
 
 def get_interface(title="", interface=None, env=None):
+    tried_tui = False
     try:
-        if interface == "tui":  # undocumented feature
+        if interface == "tui":
             interface = __getattr__("TuiInterface")
-        elif interface == "gui":  # undocumented feature
+            tried_tui = True
+        elif interface == "gui":
             interface = __getattr__("GuiInterface")
         if interface is None:
             interface = __getattr__("GuiInterface") or __getattr__("TuiInterface")
-        interface = interface(title, env)
+        return interface(title, env)
     except InterfaceNotAvailable:  # Fallback to a different interface
-        interface = __getattr__("TuiInterface")(title, env)
-    return interface
+        pass
+    if not tried_tui:
+        try:
+            return __getattr__("TuiInterface")(title, env)
+        except InterfaceNotAvailable:
+            # Even though TUI is able to claim a non-interactive terminal,
+            # ex. when doing a cron job, a terminal cannot be made interactive.
+            pass
+    return Mininterface(title, env)
 
 
 __all__ = ["GuiInterface", "TuiInterface", "TextInterface", "TextualInterface", "TkInterface"]