+++ docs/Changelog.md
@@ -2,6 +2,7 @@
 
 ## 0.7.1
 * GUI scrollbars if window is bigger than the screen
+* [non-interactive][mininterface.Mininterface.__enter__] session support
 
 ## 0.7.0 (2024-11-08)
 * hidden [`--integrate-to-system`](Overview.md#bash-completion) argument